refactor: 수익률 계산 시 BigDecimal자료형을 활용해 계산하도록 변경

+++ src/main/java/lotto/model/LottoPurchaseInformation.java
@@ -1,5 +1,7 @@
 package lotto.model;
 
+import java.math.BigDecimal;
+
 public final class LottoPurchaseInformation {
 	private static final String INSUFFICIENT_ERROR = "금액이 부족합니다.";
 	private static final int MIN_AUTO_LOTTO_COUNT = 0;
@@ -23,7 +25,7 @@ public final class LottoPurchaseInformation {
 		return purchaseAmount.getAutoLottoCount(manualLottoCount);
 	}
 
-	public double profitRate(int totalPrize) {
+	public BigDecimal profitRate(BigDecimal totalPrize) {
 		return purchaseAmount.calculateProfitRate(totalPrize);
 	}
 }
